ocr/requirement.txt
astroid==2.3.1
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
docutils==0.15.2
idna==2.8
isort==4.3.21
Kivy==1.11.1
kivy-deps.glew==0.1.12
kivy-deps.gstreamer==0.1.17
kivy-deps.sdl2==0.1.22
Kivy-Garden==0.1.4
lazy-object-proxy==1.4.2
mccabe==0.6.1
nltk==3.4.5
Pillow==6.1.0
Pygments==2.4.2
pylint==2.4.2
pyperclip==1.7.0
pypiwin32==223
pytesseract==0.3.0
pywin32==225
requests==2.22.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2